requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
et-xmlfile==1.1.0
fastapi==0.75.0
fastapi-jwt-auth==0.5.0
flake8==4.0.1
flake8-docstrings==1.6.0
glob2==0.7
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
iniconfig==1.1.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
parse==1.19.0
parse-type==0.6.0
passlib==1.7.4
pluggy==1.0.0
prettyconf==2.2.1
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pydantic==1.9.0
pydocstyle==6.1.1
pyflakes==2.4.0
PyJWT==1.7.1
pyparsing==3.0.7
pytest==7.1.1
pytest-bdd==5.0.0
python-multipart==0.0.5
requests==2.27.1
rsa==4.8
sentry-sdk==1.5.12
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.32
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
tomli==2.0.1
typing-extensions==4.1.1
unicorn==1.0.3
urllib3==1.26.9
uvicorn==0.17.6
zipp==3.7.0
asyncpg==0.25.0
pandas~=1.4.3
aioredis==2.0.1
aiohttp==3.8.1
contextvars==2.4
kafka-python==2.0.2
weasyprint==52.5
plotly==5.13.1
kaleido==0.2.1
Jinja2==3.1.2
Pillow==9.5.0
oci==2.101.0
